moving graphics part to rtrp module
