+++ cyfra-e2e-test/src/test/scala/io/computenode/cyfra/e2e/vulkan/SequenceExecutorTest.scala
@@ -1,7 +1,6 @@
 package io.computenode.cyfra.e2e.vulkan
 
-import io.computenode.cyfra.vulkan.pipeline.compute.{Binding, ComputePipeline, InputBufferSize, LayoutInfo, LayoutSet, Shader}
-import io.computenode.cyfra.vulkan.pipeline.LayoutInfo
+import io.computenode.cyfra.vulkan.compute.{Binding, ComputePipeline, InputBufferSize, LayoutInfo, LayoutSet, Shader}
 import io.computenode.cyfra.vulkan.executor.BufferAction.{LoadFrom, LoadTo}
 import io.computenode.cyfra.vulkan.executor.SequenceExecutor
 import io.computenode.cyfra.vulkan.executor.SequenceExecutor.{ComputationSequence, Compute, Dependency, LayoutLocation}
