+++ cyfra-vulkan/src/main/scala/io/computenode/cyfra/vulkan/compute/ComputePipeline.scala
@@ -1,7 +1,6 @@
-package io.computenode.cyfra.vulkan.pipeline.compute
+package io.computenode.cyfra.vulkan.compute
 
 import io.computenode.cyfra.vulkan.VulkanContext
-import io.computenode.cyfra.vulkan.pipeline.LayoutInfo
 import io.computenode.cyfra.vulkan.core.Device
 import io.computenode.cyfra.vulkan.util.Util.{check, pushStack}
 import io.computenode.cyfra.vulkan.util.VulkanObjectHandle
