+++ cyfra-vulkan/src/main/scala/io/computenode/cyfra/vulkan/compute/LayoutInfo.scala
@@ -1,4 +1,4 @@
-package io.computenode.cyfra.vulkan.pipeline
+package io.computenode.cyfra.vulkan.compute
 
 /** @author
   *   MarconZet Created 25.04.2020
